transpile: whole-function inference; an unsettled subject chooses nothing, poisoning by the exact set, the arm in its scope, kinds and joins, one representation only by value
Six items of 2026-09-08-addenda-inference-3.md's FINAL ORDER for the sixth
agent (report 2026-09-09-report-inference-6.md, progress log
2026-09-09-progress-inference-6.md), over d136899: the two mirror-image
regressions the reviews of 2273fd4 found, and the false readings behind the
budget's four contradiction rows.

While a method candidate's subject is an unsettled unknown the engine chooses
nothing at any deref depth: the walk stops at a candidate whose obligation is
unsettled, and a resolution that saw a second answer deeper is contested, its
result held at a site-stable unknown until the solve says which method it is
(a candidate a fact rules out is still overtaken). That rule needed a unit
struct written as a bare path to be typed as a value. A contradiction poisons
exactly the variables unification walked through; the fourth commit's
receiver-based provenance is retracted with it. A match arm is typed in its
own pattern scope, a range's endpoints are read by the ordinary rules, and a
projection may read through an impl whose bounds are merely unsettled as a
second pass: the budget's four "constrained to be the same type" rows are
gone because the readings are right. Kinds are merged between two variables
(a kind mismatch reported even while both stand for nothing); comparison and
boolean operands, branch joins, the negated literal and an integer index as
usize are constraints. Two types are one representation only where the
emitted code writes nothing between them, by value; through a reference
every deref hop stands, which needed a tuple struct's own name typed as its
constructor. A bound projects through the type as written and no further.

Contract choices 4.1-4.9 of the report accepted. Built, measured and taken
out: the callee's inferred bounds validated at the call, which works on its
own shapes but filed sixty rows on a corpus that compiles because
Probe::holds answers no where Rust answers yes (a reference subject against
a trait whose reflexive blanket covers it; a crate impl that exists); what is
owed is a holds that can be reported on. A golden for the match-arm double
drop could not fail at d136899, so that cost is recorded as not reproduced
and the instrument is the type-level test.

Budget: core 1523 -> 1436, signals 103 -> 99, storage-indexeddb 259 -> 255;
"constrained to be the same type" 4 -> 0; "the subject is a type the engine
has not settled" 48 -> 0 in core; untyped bindings core 187 at 127 -> 164 at
116, signals 8 at 6 -> 5 at 4. Eleven corpus files emit differently, each
hunk a now-typed local gaining a release, a closure typing its capture, or a
cast that stopped truncating. Harness 1031 -> 1052 across fifteen targets, zero
warnings; proto 301/1, ankql 422/0, base 535/0, zero leak reports and zero
fatal markers. Not landed: the await family, the settled-width literal and
the remaining reports, the fifth agent's leftovers, the hygiene sweep.
Verified in the quiet worktree and on main after the merge.

+++ transpile/goldens/match_catch_all/expected.ts
@@ -174,7 +174,7 @@ export function letInit(cause: Cause): number {
     Missing: () => 2,
     Other: () => 2,
   });
-  return checkedAdd(picked, 1, 'i32');
+  return checkedAdd(picked, 1, 'usize');
 }
 
 export function asArgument(cause: Cause): number {
